extensions: ip6t_{S,D}NAT: multiple to-dst/to-src arguments not reported
This patch is fixing the detection of multiple '--to-destination' in a
DNAT rule and '--to-source' in SNAT rule for IPv6.  Currently, when
defining multiple values for these, only the last will be used and
others ignored silently.

The checks for (cb->xflags & F_X_TO_[DEST/SRC]) always fails because the
flags are never set before.  It seems to be a copy-paste artefact since
introduction of the IPv6 DNAT/SNAT extensions based on IPv4 code.

I also removed the kernel_version checks because they seem useless.
Extensions for IPv6 DNAT/SNAT are using xt_target with revision 1.  That
seems only added since kernel version 3.7-rc1 and therefore the check
for > v2.6.10 will always return true.  The check is probably also
coming from the IPv4 copy-paste.

Add tests to cover this too, including the IPv4 side.

Signed-off-by: Thierry Du Tre <thierry@dtsystems.be>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>