xtables-compat-multi.c: Allow symlink of ebtables
This patch allows one to force a subsystem that one does not wish to modify
(e.g. libvirt) to use the ebtables compatibility layer.

ebtables-compat was already a symlink to xtables-compat-multi but ebtables was a
stand-alone program. So one could move it out of the way before making the
symlink as below:

 lrwxrwxrwx 1 root root       20 Feb 24 11:03 ebtables -> xtables-compat-multi
 -rwxr-xr-x 1 root root    75176 Feb 24 11:03 ebtables.orig

With this patch, kernel modules ebtable_filter & ebtables are no longer loaded.

Signed-off-by: Duncan Roe <duncan_roe@optusnet.com.au>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>