tests: shell: Fix for 'make distcheck'
The target performs a "VPATH build", so built binaries are not put into
the same directory tree as the test script itself. For lack of a better
way to detect this, assume $PWD in this situation remains being the
build tree's TLD and check if binaries are present in there.

Signed-off-by: Phil Sutter <phil@nwl.cc>